plan_and_reason_mcp/requirements_abicopilot.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiomysql==0.2.0
aioodbc==0.5.0
aiosignal==1.3.1
alembic==1.11.1
altair==5.1.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asn1crypto==1.5.1
astroid==2.15.6
asttokens==3.0.0
async-lru==2.0.4
async-timeout==4.0.3
asyncpg==0.29.0
atomicwrites==1.4.1
atpublic==4.0
attrs==23.1.0
autoflake==1.4
azure-identity==1.15.0
azure-storage-blob==12.19.1
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
backports.functools-lru-cache==1.6.5
bandit==1.7.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==4.2.0
bleach==6.2.0
blis==0.7.11
boto3==1.34.61
botocore==1.34.61
build==1.0.3
cachetools==5.3.1
catalogue==2.0.10
cattrs==23.2.3
celery==5.3.6
certifi==2025.1.31
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
clickhouse-connect==0.6.11
cloudpathlib==0.16.0
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
comtypes==1.4.10
confection==0.1.3
configs==3.0.3
coverage==7.3.2
cryptography==41.0.7
cssselect==1.2.0
curl_cffi==0.6.2
cymem==2.0.8
databases==0.4.2
databricks-sql-connector==2.0.2
dataclasses-json==0.6.7
DBUtils==3.0.3
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
docarray==0.40.0
docker-pycreds==0.4.0
docstring_parser==0.16
docx2txt==0.8
duckdb==0.8.1
duckduckgo_search==5.3.0
ecdsa==0.18.0
editdistance==0.8.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.104.1
FastAPI-SQLAlchemy==0.2.1
fastjsonschema==2.21.1
filelock==3.12.3
filetype==1.2.0
finnhub-python==2.4.23
flatbuffers==24.3.25
flower==2.0.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2023.9.0
fuzzywuzzy==0.18.0
geographiclib==2.0
geopy==2.4.1
ghp-import==2.1.0
gitdb==4.0.10
GitPython==3.1.40
google-ai-generativelanguage==0.4.0
google-api-core==2.11.1
google-api-python-client==2.165.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.42.1
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.14.0
google-crc32c==1.7.0
google-genai==1.10.0
google-generativeai==0.3.2
google-resumable-media==2.7.2
googleapis-common-protos==1.60.0
greenlet==3.1.1
griffe==0.38.1
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
humanfriendly==10.0
humanize==4.9.0
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.34.0
ipywidgets==8.1.5
isodate==0.6.1
isoduration==20.11.0
isort==5.12.0
jedi==0.19.2
jellyfish==1.0.3
Jinja2==3.1.2
jiter==0.9.0
jmespath==1.0.1
joblib==1.2.0
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.1.1
jupyter_client==8.6.3
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kombu==5.3.4
langchain==0.3.21
langchain-community==0.0.38
langchain-core==0.3.49
langchain-experimental==0.0.50
langchain-google-genai==0.0.5
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langchainhub==0.1.15
langcodes==3.3.0
langgraph==0.0.24
langsmith==0.1.147
lark==1.1.9
lazy-object-proxy==1.9.0
Levenshtein==0.25.0
lxml==4.9.3
lz4==4.3.2
Mako==1.2.4
Markdown==3.5.1
markdown-it-py==3.0.0
markdown2==2.4.10
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mcp==1.6.0
mdurl==0.1.2
mergedeep==1.3.4
mike==1.1.2
mistune==3.1.3
mkdocs==1.5.2
mkdocs-autorefs==0.4.1
mkdocs-gen-files==0.5.0
mkdocs-glightbox==0.3.4
mkdocs-literate-nav==0.6.0
mkdocs-material==9.2.6
mkdocs-material-extensions==1.1.1
mkdocs-section-index==0.3.5
mkdocs-video==1.5.0
mkdocstrings==0.22.0
mkdocstrings-python==1.6.0
monotonic==1.6
mpmath==1.3.0
msal==1.28.0
msal-extensions==1.0.0
msg-parser==1.2.0
multidict==6.2.0
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.1
newsapi-python==0.2.7
nltk==3.9.1
node==1.2.2
notebook==7.3.3
notebook_shim==0.2.4
numexpr==2.8.5
numpy==1.26.4
oauthlib==3.2.2
odict==1.9.0
olefile==0.46
openai==1.69.0
opencensus==0.11.2
opencensus-context==0.1.3
opencensus-ext-azure==1.1.9
openmeteo_requests==1.2.0
openmeteo_sdk==1.11.4
openpyxl==3.1.2
orjson==3.10.15
overrides==7.4.0
packaging==23.2
paginate==0.5.6
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pbr==5.11.1
pdf2image==1.17.0
pdfminer.six==20221105
pgeocode==0.4.1
pickleshare==0.7.5
pillow==11.1.0
pip==24.0
platformdirs==2.6.2
pluggy==1.3.0
plumber==1.7
portalocker==2.7.0
posthog==3.0.2
pre-commit-hooks==4.4.0
preshed==3.0.9
prometheus_client==0.20.0
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==4.24.2
psutil==7.0.0
psycopg2==2.9.6
psycopg2-binary==2.9.9
pulsar-client==3.3.0
pure_eval==0.2.3
py==1.11.0
pyarrow==15.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.11.0
pydantic_core==2.33.0
pydantic-settings==2.8.1
pydeck==0.8.1b0
Pygments==2.19.1
PyJWT==2.8.0
pylint==2.17.5
pymdown-extensions==10.5
Pympler==1.0.1
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
PyMySQL==1.1.0
pyOpenSSL==23.3.0
pypandoc==1.11
pyparsing==3.1.1
PyPDF2==3.0.1
pyphen==0.14.0
pyproject_hooks==1.0.0
pyquery==2.0.0
pyreadline3==3.4.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==3.3.0
python-Levenshtein==0.25.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywinauto==0.6.9
pywinpty==2.0.15
PyYAML==6.0.2
pyyaml_env_tag==0.1
pyzmq==26.3.0
rake-nltk==1.0.6
rapidfuzz==3.6.1
readtime==3.0.0
referencing==0.30.2
regex==2024.11.6
requests==2.31.0
requests-cache==1.2.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry-requests==2.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.10.2
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
ruff==0.0.284
s3transfer==0.10.0
safetensors==0.4.3
scikit-learn==1.2.2
scipy==1.11.2
segtok==1.5.11
Send2Trash==1.8.3
sentry-sdk==1.36.0
setproctitle==1.3.3
setuptools==69.5.1
shapely==2.0.7
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.0
sniffio==1.3.1
snowflake==0.6.0
snowflake-connector-python==3.7.0
snowflake.core==0.6.0
snowflake._legacy==0.5.0
snowflake-snowpark-python==1.13.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.39
sqlmodel==0.0.16
srsly==2.4.8
sse-starlette==2.2.1
stack-data==0.6.2
starlette==0.46.1
stevedore==5.1.0
sympy==1.12
tabulate==0.9.0
tenacity==8.5.0
terminado==0.18.1
textdistance==4.6.1
textstat==0.7.3
thinc==8.2.1
threadpoolctl==3.2.0
thrift==0.16.0
tiktoken==0.9.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
torch==2.0.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
types-python-dateutil==2.9.0.20241206
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing_extensions==4.12.2
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.1
tzlocal==4.3.1
u==1.0.4
unstructured==0.8.1
uri-template==1.3.0
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==2.3.0
uv==0.6.14
uvicorn==0.24.0
validators==0.22.0
verspec==0.1.0
vine==5.1.0
wandb==0.16.0
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.13
weasel==0.3.4
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.43.0
widgetsnbextension==4.0.13
wrapt==1.15.0
xlrd==2.0.1
XlsxWriter==3.0.4
yake==0.4.8
yarl==1.18.3
zipp==3.21.0
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.1
zope.event==5.0
zope.hookable==7.0
zope.interface==7.2
zope.lifecycleevent==5.1
zope.proxy==6.1
zstandard==0.23.0
